Refactored a lot of stuff.  BLAS tests and utility tests.  Revived RNG stuff (not doing anything with it).

+++ minibench/flop-word.c
@@ -24,15 +24,15 @@ int main() {
     CU_ERROR_CHECK(cuDeviceGetAttribute(&memoryClockRate, CU_DEVICE_ATTRIBUTE_MEMORY_CLOCK_RATE, device));
     CU_ERROR_CHECK(cuDeviceGetAttribute(&globalMemoryBusWidth, CU_DEVICE_ATTRIBUTE_GLOBAL_MEMORY_BUS_WIDTH, device));
 
-    unsigned long bandwidth = memoryClockRate * 2000ul * globalMemoryBusWidth / CHAR_BIT;
-    unsigned long flops = (major == 2 ? (minor == 0 ? 32ul : 48ul) : 8ul) * multiProcessorCount * clockRate * 2000ul;
+    long bandwidth = memoryClockRate * 2000 * globalMemoryBusWidth / CHAR_BIT;
+    long flops = (major == 2 ? (minor == 0 ? 32 : 48) : 8) * multiProcessorCount * clockRate * 2000;
 
-    fprintf(stdout, "Device %d: %3.2f GFlops/s single precision fmad, %3.2f GB/s bandwidth, flop:word %2.2f\n", d, flops * 1.e-9, bandwidth / (double)(1 << 30), flops / (bandwidth / 4.0));
+    fprintf(stdout, "Device %d: %3.2f GFlops/s single precision fmad, %3.2f GB/s bandwidth, flop:word %2.2f\n", d, (double)flops * 1.e-9, (double)bandwidth / (double)(1 << 30), (double)flops / ((double)bandwidth / 4.0));
 
     if (major >= 1 && minor >= 2) {
-      flops = (major == 2 ? (minor == 0 ? 16ul : 4ul) : 1ul) * multiProcessorCount * clockRate * 2000ul;
+      flops = (major == 2 ? (minor == 0 ? 16 : 4) : 1) * multiProcessorCount * clockRate * 2000;
 
-      fprintf(stdout, "           %3.2f GFlops/s double precision dmad, %3.2f GB/s bandwidth, flop:word %2.2f\n", flops * 1.e-9, bandwidth / (double)(1 << 30), flops / (bandwidth / 8.0));
+      fprintf(stdout, "           %3.2f GFlops/s double precision dmad, %3.2f GB/s bandwidth, flop:word %2.2f\n", (double)flops * 1.e-9, (double)bandwidth / (double)(1 << 30), (double)flops / ((double)bandwidth / 8.0));
     }
   }
